+++ Source/DECHashInterface.pas
@@ -269,17 +269,17 @@ type
     /// <summary>
     ///   Returns the number of bits the final byte of the message consists of
     /// </summary>
-    function GetFinalByteLength: UInt16;
+    function GetFinalByteLength: UInt8;
     /// <summary>
     ///   Sets the number of bits the final byte of the message consists of
     /// </summary>
-    procedure SetFinalByteLength(const Value: UInt16);
+    procedure SetFinalByteLength(const Value: UInt8);
 
     /// <summary>
     ///   Setting this to a number of bits allows to process messages which have
     ///   a length which is not a exact multiple of bytes.
     /// </summary>
-    property FinalBitLength : UInt16
+    property FinalBitLength : UInt8
       read   GetFinalByteLength
       write  SetFinalByteLength;
   end;
